[C] - FINALLY DID IT!!!!!!!!!!!!!!!!!!!!!!
I decided to delete the user from the online redis set only if he/she logs out of the page completely. Also, I plan on deleting the code in the client "disconnect" event.

Maybe I'll add some more functionality to track if the user closes the tab (but doesn't log out). In this case, I'll delete the user from the set (maybe)